Add new features, improve type safety, and update UI
- Added debug print statements to `CornerAssetPlacer` methods.
- Replaced implicit type conversions with `static_cast<int>` for better type safety.
- Commented out error-checking code in `getPolylineDistance`.
- Added comments and warning messages in `directionOfDrawing` functions.
- Adjusted insertion point calculation in `PlaceColumn`.
- Added new build configuration `PERI|x64` and updated project files.
- Included new source files `Props.cpp` and `Props.h`.
- Added new menu macro and ribbon command button for "Push Pull Props".
- Updated `centerIndex` calculation in `PlaceBracket::placeBrackets`.
- Removed redundant block position calculation in `PlaceColumn`.
- Added new include directive and class `CBrxApp` in `acrxEntryPoint.cpp`.
- Enhanced `directionOfDrawing3` function in `TiePlacer.cpp`.
- Added new configuration entries in `config.txt`.
- Implemented `placeInsideCornerPostAndPanels` and `placeAssetsAtCorners` in `InsideCornerAssetPlacer.cpp`.
- Added new PNG image file `props.png`.
- Introduced `PlaceProps` class with methods for asset placement and annotation.

+++ DefineHeight.cpp
@@ -12,7 +12,7 @@ void DefineHeight::defineHeight() {
     ads_printf(_T("\nEnter the height of the structure (mm): "));
     if (acedGetReal(NULL, &height) == RTNORM) {
         acutPrintf(_T("\nHeight defined as: %lfmm"), height);
-        globalVarHeight = height; // Convert 
+        globalVarHeight = static_cast<int>(height); // Convert 
         // Store or use the height as needed in your application
         // For example, save it to a global variable or use it directly
     }
